% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition a dataset}
\usage{
partition(df, p = 0.75, val = FALSE)
}
\arguments{
\item{df}{a data.frame or matrix object}

\item{p}{a number between 0 and 1}
}
\value{
a list containing training set and testing set
}
\description{
Partitions a single dataset into separate training and testing sets
}
\examples{
Bpartition(iris, .7)
}
