% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.trainedKnn}
\alias{predict.trainedKnn}
\title{Predict from knn models.}
\usage{
\method{predict}{trainedKnn}(trainedModel, newData, dfOut = TRUE)
}
\arguments{
\item{trainedModel}{A trainedKnn object created by train().}

\item{newData}{A data.frame with missing target values.}

\item{dfOut}{A logical. If TRUE, the predictions will be returned as a complete data.frame. If FALSE,
the predictions will be returned as a vector.}
}
\value{
Predictions for newData's target values represented as either a data.frame or vector.
}
\description{
Predicts values using a trainedKnn model.
}
\details{
newData should be formatted identical to the model's training set. Column names
and order should be identical. All columns present in training must be included. The number of rows may differ.
Also, prediction will work even if newData's target values are not missing.
}
