% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.trainedKproto}
\alias{predict.trainedKproto}
\title{Assign cluster labels}
\usage{
\method{predict}{trainedKproto}(trainedModel, newData, dfOut = TRUE)
}
\arguments{
\item{trainedModel}{A trainedKproto object created by train().}

\item{newData}{A data.frame}

\item{dfOut}{A logical. If TRUE, the predictions will be returned as a complete data.frame. If FALSE,
the predictions will be returned as a vector.}
}
\value{
The cluster labels as either a data.frame or vector
}
\description{
Assigns each observation to its most similar prototype/cluster.
}
\details{
newData should be formatted identical to the model's training set. Column names
and order should be identical. All columns present in training must be included. The number of rows may differ.
If a data.frame is returned, the labels are assigned to a new column named .Cluster
}
