% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProc.R
\name{preProc}
\alias{preProc}
\title{Preprocess data}
\usage{
preProc(
  data,
  nBins = NULL,
  binCols = NULL,
  scale = FALSE,
  target = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{data}{A tabular style data object. Ideally a data.frame. If not a data.frame,
conversion to data.frame will be attempted.}

\item{nBins}{The number of bins to use when binning. Must be a positive integer.}

\item{binCols}{A numeric specifying which columns to bin. If binCols is null when nBins has a value,
all numeric columns will be binned}

\item{scale}{A logical. If true, all numeric columns will be scaled between 0 and 1.}

\item{na.rm}{A logical. Drops rows containing NA if TRUE.}
}
\value{
A processed data.frame (may have additional attributes).
}
\description{
Preprocesses data for use with babyCaret::train().
}
\details{
If both scaling and binning are being done, binned columns will not be scaled
prior to binning. If binCols is null when nBins has a value,
all numeric columns will be binned. if binCols have been specified but nBins is none,
nBins will default to 3 after warning.

It is recommended that matrices be cast to data.frames PRIOR to preprocessing.
However, if passing a matrix to data, numeric matrices will retain all columns as numeric
a data.frame. Character matrices will be cast to a data.frame consisting entirely
of factor columns.

Attributes will be added to the output if the data is binned or scaled. For binning
These will be a logical vector, "binCols" which is true for each column
That was binned, and a list of numeric vectors "quants" which contain the quantiles
used to define the bins for a specific column. The list has as many entries as
columns that were binned.

For scaling, "scaleCols" is a logical vector TRUE for each column that was scaled.
"mins" contain the minimum for each column that was scaled, and "maxes" contain
the maximum cor each column that was scaled.

These attributes are used by matchProc() in order to process additional datasets
by the same metrics. They can also be accessed using attr() or attributes().
}
