% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTune.R
\name{reTune}
\alias{reTune}
\title{Change a tune object's parameters}
\usage{
reTune(myTune, ...)
}
\arguments{
\item{myTune}{A tune object (knnTune, aprioriTune, etc.).}

\item{...}{Named arguments with assigned values.}
}
\value{
A tune object
}
\description{
Changes any tune object's parameters. Unchanged parameters remain the same.
}
\details{
Parameters not explicitly changed will remain the same. Arguments must be explicitly named as shown in the examples.
}
