% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Test model performance}
\usage{
score(trainedModel, testSet)
}
\arguments{
\item{trainedModel}{A classification or regression model trained by babyCaret. Currently,
only trainedKnn and trainedTree models are supported.}

\item{testSet}{A data.frame. Must match the format of the training set and contain
target values.}
}
\value{
the predicton error as a Mape object for regression models and a MisClass object for classification models.
These objects are numeric with a special print method.
}
\description{
Calculates an error score using a model and testing set
}
\details{
Mean absolute percentage error is computed for regression models. Percent misclassified
is computed for classification models.
}
