% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTune.R
\name{setApriori}
\alias{setApriori}
\title{Create a tuneApriori object.}
\usage{
tuneApriori(minSup = 0.1, minConf = 0.8, maxLen = 10)
}
\arguments{
\item{minSup}{A number between 0 and 1. Sets the support at which a frequent itemset will no longer be expanded.}

\item{minConf}{A number between 0 and 1. Sets the minimum confidence a rule must have to be displayed to the user.}

\item{maxLen}{A positive integer. Sets the maximum amount of items a rule may contain.}
}
\value{
A tuneApriori object.
}
\description{
Creates an object which holds user parameters for the Apriori algorithm.
}
