% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTune.R
\name{setKproto}
\alias{setKproto}
\title{Create a tuneKproto object.}
\usage{
tuneKproto(k)
}
\arguments{
\item{k}{A positive, non-zero integer. Sets the number of clusters.}

\item{lambda}{A positive number that determines the importance of categorical attributes.}

\item{nStart}{Sets how many times the algorithm will be computed.}
}
\value{
A tuneKproto object.
}
\description{
Creates an object which holds user parameters for the K-prototypes algorithm.
}
