% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train.data.frame}
\alias{train.data.frame}
\title{ML training from data.frame}
\usage{
(data, method, tune)
}
\arguments{
\item{method}{A string indicating the ML algorithm to be used. See generic function.}

\item{tune}{A model-specific tune object. See generic function.}

\item{A}{data.frame containing the entire training set.}
}
\value{
A trained ML model.
}
\description{
Trains a machine learning model using only data.frame input
}
\details{
If used for classification or regression, data will be interpreted
as a model frame
}
