% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train ML Models}
\usage{
train(x, ...)

## Default S3 method:
train(x, data, method, tune)

## S3 method for class 'formula'
train(x, data, method, tune)

## S3 method for class 'data.frame'
train(x, method, tune)
}
\arguments{
\item{x}{In the default method, x is a vector containing target values.
In the formula method, x is a formula specifying the predictors and the target
attribute/feature. In the data.frame method, x is the entire training set.}

\item{data}{A data.frame object containing the training data. Must include
target values.}

\item{method}{A character object specifying the algorithm to be used: "knn",
"tree", "kproto", or "apriori"}

\item{tune}{A tune object specifying hyperparameter values. Can be set
using setTune().}
}
\value{
A trained ML model.
}
\description{
Handles training of machine learning models
}
