% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTune.R
\name{tuneKnn}
\alias{tuneKnn}
\title{Create a tuneKnn object.}
\usage{
tuneKnn(k)
}
\arguments{
\item{k}{A positive, non-zero integer. Determines the quantity of neighbors to be averaged for prediction.}

\item{p}{LAUREM IPSUM}

\item{featWeights}{LAUREM IPSUM}
}
\value{
A tuneKnn object.
}
\description{
Creates an object which holds user parameters for the K-nearest Neighbors algorithm.
}
\details{
If featWeights is given a numeric argument, the value will be saved in the output as
character data. It will be cast back to a numeric vector before prediction.
}
